Refactor CLI into subpackage
