+++ calkit/cli/__init__.py
@@ -1,0 +1,6 @@
+from .core import *
+
+
+def run() -> None:
+    from .main import app
+    app()
